requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional but useful for agent development
requests>=2.31.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced agent frameworks (uncomment if needed)
# langchain>=0.1.0
# langgraph>=0.1.0 